Fix parsing of artist members
The ID of artist members used to be duplicated in the dumps, having its
own dedicated tag as well as being an attribute of the member name. We
were using the dedicated tag, which was removed in 2025. Now we use the
attribute, which should work with both old and new dumps.